# docs/dashboard-ai.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Flexprice API",
    "description": "Flexprice API Service",
    "contact": {
      "name": "API Support"
    },
    "license": {
      "name": "AGPL-3.0",
      "url": "https://www.gnu.org/licenses/agpl-3.0.html"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://us.api.flexprice.io/v1",
      "description": "US Region"
    },
    {
      "url": "https://api.cloud.flexprice.io/v1",
      "description": "India Region"
    }
  ],
  "paths": {
    "/ai/assistant": {
      "post": {
        "tags": [
          "AI"
        ],
        "summary": "Ask the dashboard assistant",
        "operationId": "askDashboardAssistant",
        "requestBody": {
          "description": "Conversation",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/service.AssistantRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/service.AssistantResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "x-scope": "read",
        "x-codegen-request-body-name": "request"
      }
    },
    "/ai/pricing/parse": {
      "post": {
        "tags": [
          "AI"
        ],
        "summary": "Preview pricing with OpenRouter",
        "operationId": "parsePricing",
        "requestBody": {
          "description": "Pricing description and output schema",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParseGeminiPricingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "x-scope": "read",
        "x-codegen-request-body-name": "request"
      }
    },
    "/ai/attachments": {
      "post": {
        "operationId": "uploadAIAttachment",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "File status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/service.AIAttachmentInfo"
                }
              }
            }
          }
        }
      }
    },
    "/ai/attachments/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ],
      "get": {
        "operationId": "getAIAttachment",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "File status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/service.AIAttachmentInfo"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteAIAttachment",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Removed"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ParseGeminiPricingRequest": {
        "type": "object",
        "properties": {
          "responseSchema": {
            "type": "object",
            "properties": {}
          },
          "systemPrompt": {
            "type": "string",
            "example": "You are a pricing architect..."
          },
          "userPrompt": {
            "type": "string",
            "example": "Describe pricing for..."
          }
        }
      },
      "service.AIMessage": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "tool_call_id": {
            "type": "string"
          },
          "tool_calls": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/service.AIToolCall"
            }
          }
        }
      },
      "service.AIToolCall": {
        "type": "object",
        "properties": {
          "function": {
            "type": "object",
            "properties": {
              "arguments": {
                "type": "string"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "service.AssistantRequest": {
        "type": "object",
        "properties": {
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/service.AIMessage"
            }
          },
          "attachment_ids": {
            "type": "array",
            "maxItems": 4,
            "items": {
              "type": "string"
            }
          }
        }
      },
      "service.AssistantResponse": {
        "type": "object",
        "properties": {
          "answer": {
            "type": "string"
          },
          "environment_id": {
            "type": "string"
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/service.AssistantSource"
            }
          }
        }
      },
      "service.AssistantSource": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "retrieved_at": {
            "type": "string",
            "format": "date-time"
          },
          "attachment_id": {
            "type": "string"
          }
        }
      },
      "service.AIAttachmentInfo": {
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "size",
          "status",
          "expires_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "expires_at": {
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "sections": {
            "type": "integer"
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "description": "Enter your API key in the format *x-api-key &lt;api-key&gt;**",
        "name": "x-api-key",
        "in": "header"
      }
    }
  }
}